Setup domain models and repositories
